review: encourage nit-and-approve (#2)
Summary:
When a reviewer approves a change while simultaneously leaving small
comments, the change author can merge the change after addressing those
comments, without needing to wait for explicit re-approval from the
reviewer. We call this practice "nit-and-approve". Used appropriately,
it saves time and streamlines the review process.

wchargin-branch: nit-and-approve