adding template for documentation
This is a jekyll template that will render the sourcecred docs beautifully on GitHub pages. The documentation includes search, branding for SourceCred (with some colors made brighter to look better on the page), a news feed where announcements can go, a sidebar navigation (left) with a nested list of specific links of interest, a page navigation (right) to take the user to permalinks on the page, and a top navigation for basic about, documentation, and search pages. The GitHub repository with stars is in the upper right, along with links for social media (Twitter and GitHub) on the bottom. Finally, a circleci preview is added so that a contributor does not need to render the site locally.

Signed-off-by: vsoch <vsochat@stanford.edu>