Rewrite "How SourceCred Works" Doc
I've kept the content fairly similar, but have re-worded every concept in this document to make the ideas more approachable to people who have never encountered SourceCred before, and/or have less of a techie background.  I've also added a bit of content to let the reader flow from idea to idea in a smoother way, as well as a small intro/outro. 

Did my best to keep the concepts/examples accurate, but please review with any ideas or corrections. 

Also if the audience for this doc isn't supposed to include newbs to our community/the techosphere, then this edit may not be heading in the right direction. 

Also, also, this is my first ever commit. Be nice, but also feel free to let me know of any GitHub faux pas I'm making.