Restructure categories and create placeholder pages
Setup a bunch of categories based on discussions and the Documentation Roadmap in Roam.
Not meant to be final / set in stone, just the general outline / structure as a starting
point to see if this makes sense.

Test Plan: Ensure all the categories and pages show up in the sidebar and all links are working