Revamp homepage with more links
This changes the grid of 6 "Features" of SourceCred (which are nearly a
year old and kind of confusing more than informative tbh) with a grid of
6 featured links going to major areas of interest like how to get
involved.

The "Read the Docs" button has been demoted so that it's now just one of
the six link buddies.

Also, I renamed all the logo variants, they used to be named based on
the "feature" but now they just have a numerical ordering.