Update Discord invite link
This updates the Discord invite link so that it now drops people in the
start-here channel rather than the general channel. I've also updated
existing hardcoded references to old invites to all point to the
redirect, so that we can continue to update invite links in the future.

Possible TODO: if we find that the redirect adds an extra annoying step
for the user, we could always make a variable for the discord invite and
depend on it via code. But for now I think this is fine.

Test plan: In the CI preview, test out `https://sourcecred.io/discord`.