Setup Netlify CMS to allow creating/editing blog posts
This configures Netlify CMS to allow people to create and edit pages in our docs
repo without having to know how to use git. This commit enables it for the blog
since the content there has a simpler structure, just to see if we can get this working.

Test Plan: Go to the /admin page on the deployed site to see if it renders the NetlifyCMS
 UI correctly, displays the existing blog posts, and allows creating new ones