Update distributing-grain.md
This is a document that pulls from the "Grain" concept on the sourcecred.io page, and is a document for the Beta release. I have edited and added to the Grain doc we've already worked on so that all information is covered and have made sure the tone is consistent throughout the document.