How Cred Works.md (#124)
* Update flowing-cred.md

This is a document that pulls from the "How to Earn Cred" guide and "Cred" concept on the sourcecred.io page. The docs team has expressed wanting a short list of docs created for the Beta release, as too many could be overwhelming or difficult to understand. I have merged, edited and added to the docs we've already worked on so that all information is covered and have made sure the tone is consistent throughout the document.

* Update flowing-cred.md

* Update flowing-cred.md

editing for voice and flow

* Update flowing-cred.md