Update Beta Docs “setup-guide” (#156)
* Update Beta Docs “setup-guide”

* Update Beta Docs “setup-guide”

* Update Beta Docs “setup-guide”

Co-authored-by: LB Strobbe <63744434+LexBirdie@users.noreply.github.com>

Mostly re-wrote the intro content of this doc to make if feel a bit smoother and direct. Hopefully the content is still accurate and I'll lean on those who edit this after me to double check that. 

I also left some invisible notes about the content of this doc, primarily: how technically fluent do we expect the audience of this doc to be and how does that influence the amount of explaining we do in it, as well as how can we make this into a super direct and concrete step by step guide? 

I'll have to leave most of the content creation to those who understand the setting up an instance process better than I do.